+++ src/lib/provable-types/packed.ts
@@ -10,7 +10,7 @@ import { Poseidon, packToFields } from '../hash.js';
 import { Provable } from '../provable.js';
 import { fields } from './fields.js';
 
-export { Packed };
+export { Packed, Hashed };
 
 /**
  * Packed<T> is a "packed" representation of any type T.
@@ -136,4 +136,102 @@ function countFields(input: HashInput) {
   if (pendingBits > 0) n++;
 
   return n;
+}
+
+/**
+ * Hashed<T> represents a type T by its hash.
+ *
+ * Since a hash is only a single field element, this can be more efficient in provable code
+ * where the number of constraints depends on the number of field elements per value.
+ *
+ * For example, `Provable.if(bool, x, y)` takes O(n) constraints, where n is the number of field
+ * elements in x and y. With Hashed, this is reduced to O(1).
+ *
+ * The downside is that you will pay the overhead of hashing your values, so it helps to experiment
+ * in which parts of your code a hashed representation is beneficial.
+ *
+ * Usage:
+ *
+ * ```ts
+ * // define a hashed type from a type
+ * let HashedType = Hashed.create(MyType);
+ *
+ * // hash a value
+ * let hashed = HashedType.hash(value);
+ *
+ * // ... operations on hashes, more efficient than on plain values ...
+ *
+ * // unhash to get the original value
+ * let value = hashed.unhash();
+ * ```
+ */
+class Hashed<T> {
+  hash: Field;
+  value: Unconstrained<T>;
+
+  /**
+   * Create a hashed representation of `type`. You can then use `HashedType.hash(x)` to wrap a value in a `Hashed`.
+   */
+  static create<T>(type: ProvableExtended<T>): typeof Hashed<T> {
+    return class Hashed_ extends Hashed<T> {
+      static _innerProvable = type;
+      static _provable = provableFromClass(Hashed_, {
+        hash: Field,
+        value: Unconstrained.provable,
+      }) as ProvableHashable<Hashed<T>>;
+    };
+  }
+
+  constructor(hash: Field, value: Unconstrained<T>) {
+    this.hash = hash;
+    this.value = value;
+  }
+
+  /**
+   * Wrap a value, and represent it by its hash in provable code.
+   */
+  static hash<T>(x: T): Hashed<T> {
+    let input = this.innerProvable.toInput(x);
+    let packed = packToFields(input);
+    let hash = Poseidon.hash(packed);
+    return new this(hash, Unconstrained.from(x));
+  }
+
+  /**
+   * Unwrap a value from its hashed variant.
+   */
+  unhash(): T {
+    let value = Provable.witness(this.Constructor.innerProvable, () =>
+      this.value.get()
+    );
+
+    // prove that the value hashes to the hash
+    let input = this.Constructor.innerProvable.toInput(value);
+    let packed = packToFields(input);
+    let hash = Poseidon.hash(packed);
+    this.hash.assertEquals(hash);
+
+    return value;
+  }
+
+  toFields(): Field[] {
+    return [this.hash];
+  }
+
+  // dynamic subclassing infra
+  static _provable: ProvableHashable<Hashed<any>> | undefined;
+  static _innerProvable: ProvableExtended<any> | undefined;
+
+  get Constructor(): typeof Hashed {
+    return this.constructor as typeof Hashed;
+  }
+
+  static get provable(): ProvableHashable<Hashed<any>> {
+    assert(this._provable !== undefined, 'Hashed not initialized');
+    return this._provable;
+  }
+  static get innerProvable(): ProvableExtended<any> {
+    assert(this._innerProvable !== undefined, 'Hashed not initialized');
+    return this._innerProvable;
+  }
 }